+++ src/Data/Singletons/TypeLits/Internal.hs
@@ -199,28 +199,3 @@ sa %<> sb =
 infixr 6 %<>
 
 $(genDefunSymbols [''(<>)])
-
-------------------------------------------------------------
--- TypeLits singleton non-singleton instances
-------------------------------------------------------------
-
--- Thanks to @cumber on #179
-
-instance Show (SNat n) where
-  showsPrec p n@SNat
-    = showParen (p > atPrec)
-      ( showString "SNat @"
-        . showsPrec (atPrec + 1) (TN.natVal n)
-      )
-    where atPrec = 10
-
-instance Show (SSymbol s) where
-  showsPrec p s@SSym
-    = showParen (p > atPrec)
-      ( showString "SSym @"
-        . showsPrec (atPrec + 1) (symbolVal s)
-      )
-    where atPrec = 10
-
-deriving instance Show (SomeSing Nat)
-deriving instance Show (SomeSing Symbol)